requirements.txt
numpy>=1.24.0
scikit-learn==1.3.2
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyserial>=3.5
micromlgen==1.3.6
